Rename modules
batch_status_updater -> batch_transaction_fetcher
batch_transaction_updater -> transaction_finality_updater
