+++ core/lib/node_framework/tests/ui/incorrect/01_from_context_task.rs
@@ -16,7 +16,7 @@ struct TaskA;
 #[async_trait::async_trait]
 impl Task for TaskA {
     fn id(&self) -> TaskId {
-        "batch_status_updater".into()
+        "batch_transaction_fetcher".into()
     }
 
     async fn run(self: Box<Self>, _stop_receiver: StopReceiver) -> anyhow::Result<()> {
