+++ core/lib/node_framework/tests/ui/incorrect/02_into_context_default_task.rs
@@ -17,7 +17,7 @@ struct TaskA;
 #[async_trait::async_trait]
 impl Task for TaskA {
     fn id(&self) -> TaskId {
-        "batch_status_updater".into()
+        "batch_transaction_fetcher".into()
     }
 
     async fn run(self: Box<Self>, _stop_receiver: StopReceiver) -> anyhow::Result<()> {
