+++ core/node/node_sync/src/lib.rs
@@ -1,5 +1,4 @@
-pub mod batch_status_updater;
-pub mod batch_transaction_updater;
+pub mod batch_transaction_fetcher;
 mod client;
 pub mod data_availability_fetcher;
 pub mod external_io;
@@ -12,6 +11,7 @@ pub mod sync_action;
 pub mod testonly;
 #[cfg(test)]
 mod tests;
+pub mod transaction_finality_updater;
 pub mod tree_data_fetcher;
 pub mod validate_chain_ids_task;
