+++ core/node/node_sync/src/node/mod.rs
@@ -1,18 +1,18 @@
 pub use self::{
-    batch_status_updater::BatchStatusUpdaterLayer,
-    batch_transaction_updater::BatchTransactionUpdaterLayer,
+    batch_transaction_fetcher::BatchStatusUpdaterLayer,
     data_availability_fetcher::DataAvailabilityFetcherLayer, external_io::ExternalIOLayer,
     miniblock_precommit_fetcher::MiniblockPrecommitFetcherLayer,
     resources::ActionQueueSenderResource, sync_state_updater::SyncStateUpdaterLayer,
+    transaction_finality_updater::BatchTransactionUpdaterLayer,
     tree_data_fetcher::TreeDataFetcherLayer, validate_chain_ids::ValidateChainIdsLayer,
 };
 
-mod batch_status_updater;
-mod batch_transaction_updater;
+mod batch_transaction_fetcher;
 mod data_availability_fetcher;
 mod external_io;
 mod miniblock_precommit_fetcher;
 mod resources;
 mod sync_state_updater;
+mod transaction_finality_updater;
 mod tree_data_fetcher;
 mod validate_chain_ids;
